Update and rename AddItemView.vue to ItemView.vue
